fix(ci): maintainability line limit — split oversized pr_monitor_runner part_004 tests into part_020
